+++ pytorch_lightning/loops/optimization/manual_loop.py
@@ -103,7 +103,7 @@ class ManualOptimization(Loop[_OUTPUTS_TYPE]):
 
             # manually capture logged metrics
             training_step_output = self.trainer._call_strategy_hook("training_step", *step_kwargs.values())
-            self.trainer.training_type_plugin.post_training_step()
+            self.trainer.strategy.post_training_step()
 
             del step_kwargs