+++ tests/deprecated_api/test_remove_1-8.py
@@ -250,3 +250,9 @@ def test_v1_8_0_deprecate_trainer_callback_hook_mixin():
         )
     with pytest.deprecated_call(match="was deprecated in v1.6 and will be removed in v1.8"):
         trainer.on_before_zero_grad(optimizer=optim.SGD(model.parameters(), lr=0.01, momentum=0.9))
+
+
+def test_v1_8_0_deprecated_training_type_plugin_property():
+    trainer = Trainer()
+    with pytest.deprecated_call(match="in v1.6 and will be removed in v1.8"):
+        trainer.training_type_plugin